Refactor saveTransfer
