+++ src/Dyst.ts
@@ -5,7 +5,7 @@ import { loadOrCreateTransaction } from './utils/Transactions'
 import { updateTreasuryRevenueDystTransfer } from './utils/TreasuryRevenue'
 
 import { DAO_WALLET, DAO_WALLET_PENROSE_USER_PROXY, DYSTOPIA_TRACKED_GAUGES } from './utils/Constants'
-import { addressEqualsString } from './utils'
+import { addressEqualsString, saveTransfer } from './utils'
 
 export function handleDystTransfer(event: TransferEvent): void {
   if (
@@ -18,14 +18,8 @@ export function handleDystTransfer(event: TransferEvent): void {
       event.params.from.toHexString(),
       event.params.to.toHexString(),
     ])
-    let transaction = loadOrCreateTransaction(event.transaction, event.block)
-    let entity = new Transfer(transaction.id)
-    entity.transaction = transaction.id
-    entity.timestamp = transaction.timestamp
-    entity.from = event.params.from
-    entity.to = event.params.to
-    entity.value = event.params.value
 
+    let entity = saveTransfer(event)
     //Pass entity to TreasuryRevenue
     updateTreasuryRevenueDystTransfer(entity)
     entity.save()
