+++ src/OtterClamERC20V2.ts
@@ -6,7 +6,7 @@ import { loadOrCreateTransaction } from './utils/Transactions'
 import { getClamUsdRate } from './utils/Price'
 import { DAO_WALLET, OTTOPIA_STORE, OTTO_PORTAL_MINTING } from './utils/Constants'
 import { loadOrCreateTreasuryRevenue } from './utils/TreasuryRevenue'
-import { addressEqualsString } from './utils'
+import { addressEqualsString, saveTransfer } from './utils'
 
 export function handleTransfer(event: TransferEvent): void {
   //BURN events
@@ -63,15 +63,4 @@ export function loadOrCreateTotalBurnedClamSingleton(): TotalBurnedClam {
     total.burnedValueUsd = BigDecimal.zero()
   }
   return total
-}
-
-function saveTransfer(event: TransferEvent): void {
-  let transaction = loadOrCreateTransaction(event.transaction, event.block)
-  let entity = new Transfer(transaction.id)
-  entity.from = event.params.from
-  entity.to = event.params.to
-  entity.value = event.params.value
-  entity.timestamp = transaction.timestamp
-  entity.transaction = transaction.id
-  entity.save()
 }
