+++ src/Qi.ts
@@ -3,16 +3,8 @@ import { Address, log } from '@graphprotocol/graph-ts'
 import { Transfer, TotalBribeReward } from '../generated/schema'
 import { loadOrCreateTransaction } from './utils/Transactions'
 import { updateTreasuryRevenueQiTransfer, loadOrCreateTotalBribeRewardsSingleton } from './utils/TreasuryRevenue'
-import { addressEqualsString } from './utils'
-import {
-  OTTER_DEPLOYER,
-  QI_BRIBE_REWARDS,
-  TREASURY_ADDRESS,
-  UNI_MAI_USDC_QI_INVESTMENT_PAIR,
-  UNI_QI_WMATIC_INVESTMENT_PAIR,
-} from './utils/Constants'
-import { toDecimal } from './utils/Decimals'
-import { getQiUsdRate } from './utils/Price'
+import { addressEqualsString, saveTransfer } from './utils'
+import { TREASURY_ADDRESS, UNI_MAI_USDC_QI_INVESTMENT_PAIR, UNI_QI_WMATIC_INVESTMENT_PAIR } from './utils/Constants'
 
 export function handleQiTransfer(event: TransferEvent): void {
   if (
@@ -25,14 +17,7 @@ export function handleQiTransfer(event: TransferEvent): void {
       event.params.from.toHexString(),
       event.params.to.toHexString(),
     ])
-    let transaction = loadOrCreateTransaction(event.transaction, event.block)
-    let entity = new Transfer(transaction.id)
-    entity.transaction = transaction.id
-    entity.timestamp = transaction.timestamp
-    entity.from = event.params.from
-    entity.to = event.params.to
-    entity.value = event.params.value
-    entity.save()
+    let entity = saveTransfer(event)
 
     //Pass entity to TreasuryRevenue
     updateTreasuryRevenueQiTransfer(entity)
